+++ frontend/src/containers/article/ArticleViewerContainer.js
@@ -1,16 +1,16 @@
 import React, { useEffect } from 'react';
 import ArticleViewer from '../../components/article/ArticleViewer';
 import { useDispatch, useSelector } from 'react-redux';
-import { readArticle, unloadArticle } from '../../modules/article/read/readArticleActions';
+import { readArticle, unloadArticle } from '../../modules/article/articles/articlesActions';
 import { useParams } from 'react-router-dom';
 
 const ArticleViewerContainer = () => {
 	const { articleId } = useParams();
 	const dispatch = useDispatch();
-	const { article, error, loading } = useSelector(({ article, loading }) => ({
-		article: article.article,
-		error: article.error,
-		loading: loading['article/READ_ARTICLE'],
+	const { article, error, loading } = useSelector(({ articles, loading }) => ({
+		article: articles.article,
+		error: articles.error,
+		loading: loading['articles/READ_ARTICLE'],
 	}));
 
 	useEffect(() => {
