+++ frontend/src/containers/article/MyArticlesContainer.js
@@ -1,12 +1,12 @@
 import React, { useEffect } from 'react';
 import MyArticles from '../../components/article/MyArticles';
 import { useDispatch, useSelector } from 'react-redux';
-import { readMyArticles, unloadArticles } from '../../modules/article/read/readArticleActions';
+import { readMyArticles, unloadArticles } from '../../modules/article/articles/articlesActions';
 const MyArticlesContainer = () => {
 	const dispatch = useDispatch();
-	const { articles, error, loading } = useSelector(({ article, loading }) => ({
-		articles: article.articles,
-		error: article.error,
+	const { articles, error, loading } = useSelector(({ articles, loading }) => ({
+		articles: articles.articles,
+		error: articles.error,
 		loading: loading['article/READ_MY_ARTICLES'],
 	}));
