+++ frontend/src/modules/article/articles/articlesActions.js
@@ -1,10 +1,5 @@
 import { createAction } from 'redux-actions';
-import {
-	READ_ARTICLE,
-	READ_MY_ARTICLES,
-	UNLOAD_ARTICLE,
-	UNLOAD_ARTICLES,
-} from './readArticleTypes';
+import { READ_ARTICLE, READ_MY_ARTICLES, UNLOAD_ARTICLE, UNLOAD_ARTICLES } from './articlesTypes';
 
 export const readArticle = createAction(READ_ARTICLE, (id) => id);
 export const readMyArticles = createAction(READ_MY_ARTICLES);
