+++ frontend/src/modules/index.js
@@ -5,19 +5,19 @@ import { all } from 'redux-saga/effects';
 import articleEditor from './article/editor/articleEditorReducer';
 import { articleEditorSaga } from './article/editor/articleEditorSaga';
 import user, { userSaga } from './user/user';
-import { articleSaga } from './article/read/readArticleSaga';
-import article from './article/read/readArticleReducer';
+import { articlesSaga } from './article/articles/articlesSaga';
+import articles from './article/articles/articlesReducer';
 
 const rootReducer = combineReducers({
 	auth,
 	loading,
 	user,
 	articleEditor,
-	article,
+	articles,
 });
 
 export function* rootSaga() {
-	yield all([authSaga(), userSaga(), articleEditorSaga(), articleSaga()]);
+	yield all([authSaga(), userSaga(), articleEditorSaga(), articlesSaga()]);
 }
 
 export default rootReducer;
